move conversion function to own file
